requirements.txt
django>=3.2.14
langchain>=0.2.0 
langchain-community>=0.2.0 
langchain-experimental>=0.0.60 
langchain-huggingface>=0.0.3 
langchain-google-genai>=1.0.8 
google-generative-ai>=0.7.2 
docx2txt>=0.8 
chromadb>=0.5.0 
sentence-transformers>=3.0.1
